vae develop & updates
