+++ modules/__init__.py
@@ -1,3 +1,2 @@
-from .units import *
 from .unet import *
 from .vae import *
